Fixes #18161: Add tests to ir
